requirements.txt
typer[all]==0.12.5
rich==13.9.4
pydantic==2.9.2
pyyaml==6.0.2
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1
